Changing folder structure
Instead of initializing the repo at ~/.vim/pack
and storing non-package files there, it's better
to have the repo at ~/.vim and put only packages
under /pack.

Therefore the repo was moved to the parent folder,
the submodules got added back with the new path,
and vimrc got also changed a bit.